dev_requirements.txt
CommonMark==0.6.4
python-dateutil
jinja2
rtyaml
email_validator

git+https://github.com/JoshData/commonmark-py-plaintext
git+https://github.com/if-then-fund/django-html-emailer
git+https://github.com/if-then-fund/django-database-storage

dj-database-url==0.4.1

django==1.9.8
django-allauth
jsonfield
django-bootstrap3
django-notifications-hq

selenium
